requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
pandas==2.2.2
numpy>=1.24.0
openpyxl==3.1.5
requests==2.31.0
openai>=1.0.0
ortools>=9.8.0
deap>=1.4.1
